Move extensions to new file
